cloud_functions/feature-transform/repo_ml_ready/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
pandas==2.*
google-cloud-bigquery
pandas
db-dtypes